requirements.txt
Django==5.2.*
psycopg2==2.9.*
python-dotenv==1.1.1
dj-database-url==3.0.1
